tambah nilai siswa

+++ resources/views/poin/poin.blade.php
@@ -56,8 +56,10 @@ if (Auth::user()->tahun == $tahun->tahun || Auth::user()->level == "Super_admin"
                         <th scope="col">Tanggal</th>
                         <th scope="col">Point</th>
                         <th scope="col">Tahun Ajaran</th>
-                        @if ($g->wakel == "BK" || Auth::user()->level == "Super_admin" || Auth::user()->level == "Tata_usaha")
-                        <th scope="col">Aksi</th>
+                        @if (Auth::user()->kelas == "-")
+                            @if ($g->wakel == "BK" || Auth::user()->level == "Super_admin" || Auth::user()->level == "Tata_usaha")
+                            <th scope="col">Aksi</th>
+                            @endif
                         @endif
                     </thead>
                     <tbody>
@@ -71,11 +73,13 @@ if (Auth::user()->tahun == $tahun->tahun || Auth::user()->level == "Super_admin"
                                 <td data-title="Tanggal">{{$po->tgl}}</td>
                                 <td data-title="Point">{{$po->poin}}</td>
                                 <td data-title="Tahun Ajaran">{{$po->tahun}}</td>
-                                @if ($g->wakel == "BK" || Auth::user()->level == "Super_admin" || Auth::user()->level == "Tata_usaha")
-                                <td>
-                                    <a href="edit/{{$po->id}}" class="btn btn-warning"><i class="fa-solid fa-pen-to-square"></i>Edit</a>
-                                    <a href="hapus_poin/{{$po->id}}" class="btn btn-danger" onclick="javascript: return confirm('Konfirmasi data akan dihapus');"><i class="fa-solid fa-trash"></i> Hapus</a>
-                                </td>
+                                @if (Auth::user()->kelas == "-")
+                                    @if ($g->wakel == "BK" || Auth::user()->level == "Super_admin" || Auth::user()->level == "Tata_usaha")
+                                    <td>
+                                        <a href="edit/{{$po->id}}" class="btn btn-warning"><i class="fa-solid fa-pen-to-square"></i>Edit</a>
+                                        <a href="hapus_poin/{{$po->id}}" class="btn btn-danger" onclick="javascript: return confirm('Konfirmasi data akan dihapus');"><i class="fa-solid fa-trash"></i> Hapus</a>
+                                    </td>
+                                    @endif
                                 @endif
                             </tr>
                         @endforeach
